requirements.txt
PySide6~=6.11.0
numpy~=2.4.4
pandas
qtawesome
librosa~=0.11.0
PyOpenGL~=3.1.10
pyqtgraph~=0.14.0
numba~=0.64.0
matplotlib